+++ packages/host/src/index.ts
@@ -1,2 +1,7 @@
-export { isInsideContainer, getHostLocalStorage, getHostProvider } from "./container.js";
+export {
+    isInsideContainer,
+    isInsideContainerSync,
+    getHostLocalStorage,
+    getHostProvider,
+} from "./container.js";
 export type { HostLocalStorage } from "./types.js";