+++ packages/tx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @polkadot-apps/tx
 
+## 0.3.3
+
+### Patch Changes
+
+- @polkadot-apps/keys@0.4.2
+
 ## 0.3.2
 
 ### Patch Changes